pyproject.toml
[tool.poetry]
name = "run4office"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.54.3"
chromadb = "^0.5.18"
langchain-openai = "^0.2.6"
langchain-text-splitters = "^0.3.2"
python-dotenv = "^1.0.1"
langgraph = "^0.2.45"
langchain = "^0.3.7"
uvicorn = "^0.32.0"
fastapi = "^0.115.4"
beautifulsoup4 = "^4.12.3"
selenium = "^4.26.1"
langdetect = "^1.0.9"
webdriver-manager = "^4.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
